+++ application/views/Alvin.php
@@ -1,3 +1,0 @@
-Hello alvin
-
-<a href="/test2">link text</a>