+++ application/views/Player.php
@@ -1,0 +1,4 @@
+<h1>Player Portfolio</h1>
+<a href="/">Home Page</a>
+</br>
+<a href="/stockhistory">Stock History</a>