+++ application/views/StockHistory.php
@@ -1,0 +1,4 @@
+<h1>StockHistory</h1>
+<a href="/">Home Page</a>
+</br>
+<a href="/player">Player Portfolio</a>