+++ CHANGELOG.md
@@ -16,3 +16,7 @@ All notable changes to this project will be documented in this file.
 - This CHANGELOG file to serve as a file to record all changes
 - Initiated github organization and added members
 - Edited Readme.md file
+## 0.0.2 - 2016-02-16 - 1:35 am
+### Added
+- Controllers and View structure properly setup for homepage, stock history page, and player page. 
+- Database queries implemented.